@@ -28,7 +28,7 @@
     "framer-motion": "^10.10.0",
     "js-cookie": "^3.0.1",
     "keycloak-js": "^20.0.5",
-    "next": "13.2.4",
+    "next": "^13.3.0",
     "pkce-challenge": "^3.1.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
