@@ -17,32 +17,38 @@ import { testApi } from '@/app/_resource/data/test.api';
 import createWebStorage from 'redux-persist/es/storage/createWebStorage';
 import { sidebarTreeSlice } from '@/app/_components/MainLayout/Sidebar/sidebarTree.slice';
 import { authSlice } from '@/app/_auth/data/auth.slice';
+import storage from 'redux-persist/lib/storage'
 
-const createNoopStorage = () => {
-  return {
-    getItem(_key: any) {
-      return Promise.resolve(null);
-    },
-    setItem(_key: any, value: any) {
-      return Promise.resolve(value);
-    },
-    removeItem(_key: any) {
-      return Promise.resolve();
-    },
-  };
-};
-
-const storage =
-  typeof window !== 'undefined'
-    ? createWebStorage('local')
-    : createNoopStorage();
+// Ниже код для исправления ошибки "redux-persist failed to create sync storage. falling back to noop storage"
+// const createNoopStorage = () => {
+//   return {
+//     getItem(_key: any) {
+//       return Promise.resolve(null);
+//     },
+//     setItem(_key: any, value: any) {
+//       return Promise.resolve(value);
+//     },
+//     removeItem(_key: any) {
+//       return Promise.resolve();
+//     },
+//   };
+// };
+// const storage =
+//   typeof window !== 'undefined'
+//     ? createWebStorage('local')
+//     : createNoopStorage();
+// Выше код для исправления ошибки "redux-persist failed to create sync storage. falling back to noop storage"
 
 const persistConfig = {
   key: 'root',
   storage,
   // Если используем RTK-query нужно обзяательно включить в blacklist ! ! !
-  // whitelist: ['auth'], // только это хотим сохрать в localstorage, остальное нам не нужно сохранять
-  // blacklist: ['auth'], // то что не хотим сохранять в localstorage
+  whitelist: ['auth', 'sidebarTree'], // только это хотим сохрать в localstorage, остальное нам не нужно сохранять
+  blacklist: [
+    authApi.reducerPath,
+    resourceApi.reducerPath,
+    testApi.reducerPath,
+  ], // то что не хотим сохранять в localstorage
 };
 
 const rootReducer = combineReducers({
